refactor: extract VimState lifecycle into state.ts
